scripts/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
matplotlib==3.8.2
mplfinance==0.12.10b0
